+++ puffinsystems_homepage/src/lib/components/cardCarousel/CardCarousel.svelte
@@ -1,13 +1,15 @@
 <script lang="ts">
   import type { CardContent } from "$lib/types/card";
+
   import { onMount } from "svelte";
-  import CarouselCardTemplate from "$lib/components/cardCarousel/carouselCardTemplate.svelte";
-  import CarouselButtons from "$lib/components/cardCarousel/carouselButtons.svelte";
+  import ProjectCard from "$lib/components/cardCarousel/ProjectCard.svelte";
+  import CarouselMenu from "$lib/components/cardCarousel/CarouselMenu.svelte";
 
   export let cardContentList: CardContent[];
   export let carouselName: string;
 
   let carousel: HTMLElement;
+
   let xLeft: number;
   let xWidth: number;
   let xScroll: number;
@@ -17,18 +19,17 @@
     xWidth = carousel.clientWidth;
     xScroll = carousel.scrollWidth;
   }
+
   onMount(() => {
     parseScroll();
   });
-
-  //   $: console.log({ xLeft, xWidth, xScroll });
 </script>
 
-<CarouselButtons {cardContentList} {carouselName} {xLeft} {xWidth} />
+<CarouselMenu {cardContentList} {carouselName} {xLeft} {xWidth} />
 <div class="carousel w-full" bind:this={carousel} on:scroll={parseScroll}>
   {#each cardContentList as cardContent, i}
     <div id="{carouselName}-{i + 1}" class="carousel-item w-full">
-      <CarouselCardTemplate {cardContent} />
+      <ProjectCard {cardContent} />
     </div>
   {/each}
 </div>
