+++ puffinsystems_homepage/src/lib/components/cardCarousel/CarouselMenu.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
   import type { CardContent } from "$lib/types/card";
 
+  // import CarouselMenuButton from "./CarouselMenuButton.svelte";
+
   export let cardContentList: CardContent[];
   export let carouselName: string;
   export let xLeft: number;
@@ -9,6 +11,7 @@
 
 <div class="flex gap-1 py-2">
   {#each cardContentList as cardContent, i}
+    <!-- <CarouselMenuButton {carouselName} buttonIndex={i} bind:xLeft bind:xWidth /> -->
     <a
       href="#{carouselName}-{i + 1}"
       class="btn btn-sm {xLeft >= xWidth * i && xLeft < xWidth * (i + 1)
