+++ puffinsystems_homepage/src/lib/components/cardTemplate/SideSplitCard.svelte
